+++ config/services.php
@@ -6,17 +6,22 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use Flux\VerifactuBundle\Factory\AeatResponseFactory;
 use Flux\VerifactuBundle\Factory\BreakdownDetailFactory;
+use Flux\VerifactuBundle\Factory\CancellationRecordFactory;
 use Flux\VerifactuBundle\Factory\ComputerSystemFactory;
 use Flux\VerifactuBundle\Factory\FiscalIdentifierFactory;
+use Flux\VerifactuBundle\Factory\ForeignFiscalIdentifierFactory;
 use Flux\VerifactuBundle\Factory\InvoiceIdentifierFactory;
 use Flux\VerifactuBundle\Factory\RegistrationRecordFactory;
 use Flux\VerifactuBundle\FluxVerifactuBundle;
 use Flux\VerifactuBundle\Handler\AeatClientHandler;
 use Flux\VerifactuBundle\Handler\QrCodeHandler;
+use Flux\VerifactuBundle\Handler\XmlRecordHandler;
 use Flux\VerifactuBundle\Transformer\AeatResponseTransformer;
 use Flux\VerifactuBundle\Transformer\BreakdownDetailTransformer;
+use Flux\VerifactuBundle\Transformer\CancellationRecordTransformer;
 use Flux\VerifactuBundle\Transformer\ComputerSystemTransformer;
 use Flux\VerifactuBundle\Transformer\FiscalIdentifierTransformer;
+use Flux\VerifactuBundle\Transformer\ForeignFiscalIdentifierTransformer;
 use Flux\VerifactuBundle\Transformer\InvoiceIdentifierTransformer;
 use Flux\VerifactuBundle\Transformer\RegistrationRecordTransformer;
 use Flux\VerifactuBundle\Validator\ContractsValidator;
@@ -30,6 +35,7 @@ return static function (ContainerConfigurator $container): void {
             ->args([
                 abstract_arg(FluxVerifactuBundle::AEAT_CLIENT_KEY),
                 service(RegistrationRecordFactory::class),
+                service(CancellationRecordFactory::class),
                 service(ComputerSystemFactory::class),
                 service(FiscalIdentifierFactory::class),
                 service(AeatResponseFactory::class),
@@ -44,6 +50,14 @@ return static function (ContainerConfigurator $container): void {
             ])
             ->alias(QrCodeHandler::class, 'flux_verifactu.qr_code_handler')
             ->public()
+        ->set('flux_verifactu.xml_record_handler', XmlRecordHandler::class)
+            ->args([
+                service(RegistrationRecordFactory::class),
+                service(CancellationRecordFactory::class),
+                service(ComputerSystemFactory::class),
+            ])
+            ->alias(XmlRecordHandler::class, 'flux_verifactu.xml_record_handler')
+            ->public()
         // factories
         ->set('flux_verifactu.aeat_response_factory', AeatResponseFactory::class)
             ->args([
@@ -57,6 +71,13 @@ return static function (ContainerConfigurator $container): void {
                 service(ContractsValidator::class),
             ])
             ->alias(BreakdownDetailFactory::class, 'flux_verifactu.breakdown_detail_factory')
+        ->set('flux_verifactu.cancellation_record_factory', CancellationRecordFactory::class)
+            ->args([
+                service(InvoiceIdentifierFactory::class),
+                service(CancellationRecordTransformer::class),
+                service(ContractsValidator::class),
+            ])
+            ->alias(CancellationRecordFactory::class, 'flux_verifactu.cancellation_record_factory')
         ->set('flux_verifactu.computer_system_factory', ComputerSystemFactory::class)
             ->args([
                 abstract_arg(FluxVerifactuBundle::COMPUTER_SYSTEM_CONFIG_KEY),
@@ -71,6 +92,12 @@ return static function (ContainerConfigurator $container): void {
                 service(ContractsValidator::class),
             ])
             ->alias(FiscalIdentifierFactory::class, 'flux_verifactu.fiscal_identifier_factory')
+        ->set('flux_verifactu.foreign_fiscal_identifier_factory', ForeignFiscalIdentifierFactory::class)
+            ->args([
+                service(ForeignFiscalIdentifierTransformer::class),
+                service(ContractsValidator::class),
+            ])
+            ->alias(ForeignFiscalIdentifierFactory::class, 'flux_verifactu.foreign_fiscal_identifier_factory')
         ->set('flux_verifactu.invoice_identifier_factory', InvoiceIdentifierFactory::class)
             ->args([
                 service(InvoiceIdentifierTransformer::class),
@@ -82,6 +109,7 @@ return static function (ContainerConfigurator $container): void {
                 service(InvoiceIdentifierFactory::class),
                 service(BreakdownDetailFactory::class),
                 service(FiscalIdentifierFactory::class),
+                service(ForeignFiscalIdentifierFactory::class),
                 service(RegistrationRecordTransformer::class),
                 service(ContractsValidator::class),
             ])
@@ -94,10 +122,14 @@ return static function (ContainerConfigurator $container): void {
             ->alias(AeatResponseTransformer::class, 'flux_verifactu.aeat_response_transformer')
         ->set('flux_verifactu.breakdown_detail_transformer', BreakdownDetailTransformer::class)
             ->alias(BreakdownDetailTransformer::class, 'flux_verifactu.breakdown_detail_transformer')
+        ->set('flux_verifactu.cancellation_record_transformer', CancellationRecordTransformer::class)
+            ->alias(CancellationRecordTransformer::class, 'flux_verifactu.cancellation_record_transformer')
         ->set('flux_verifactu.computer_system_transformer', ComputerSystemTransformer::class)
             ->alias(ComputerSystemTransformer::class, 'flux_verifactu.computer_system_transformer')
         ->set('flux_verifactu.fiscal_identifier_transformer', FiscalIdentifierTransformer::class)
             ->alias(FiscalIdentifierTransformer::class, 'flux_verifactu.fiscal_identifier_transformer')
+        ->set('flux_verifactu.foreign_fiscal_identifier_transformer', ForeignFiscalIdentifierTransformer::class)
+            ->alias(ForeignFiscalIdentifierTransformer::class, 'flux_verifactu.foreign_fiscal_identifier_transformer')
         ->set('flux_verifactu.invoice_identifier_transformer', InvoiceIdentifierTransformer::class)
             ->alias(InvoiceIdentifierTransformer::class, 'flux_verifactu.invoice_identifier_transformer')
         ->set('flux_verifactu.registration_record_transformer', RegistrationRecordTransformer::class)